totemnet.c: Make totemnet_initialize definition match just-changed decl.
* exec/totemnet.c (totemnet_initialize): Declare deliver_fn's msg_len
parameter to be of type "unsigned int" (not size_t) to match decl.

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2174 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ exec/totemnet.c
@@ -1676,7 +1676,7 @@ int totemnet_initialize (
 	void (*deliver_fn) (
 		void *context,
 		const void *msg,
-		size_t msg_len),
+		unsigned int msg_len),
 
 	void (*iface_change_fn) (
 		void *context,
